requerimieento finalizado

+++ src/main/java/com/reto/tecnico/model/response/OperacionResponse.java
@@ -9,9 +9,9 @@ import lombok.Setter;
 @Builder
 public class OperacionResponse {
 
-	private float monto;
+	private double monto;
 	
-	private float nuevoMonto;
+	private double nuevoMonto;
 	
 	private String  monedaDestino;
